// src/types.ts
export interface UserTypes {
  id: number;
  name: string;
  surname: string;
  patronymic: string;
  username: string;
  identification: string;
  idnumber: number;
  birthPlace: string;
  birthDate: Date;
  pincode: string;
  bloodGroupId: string;
  maritalStatus: string;
  gender: string;
  citizenship: string;
  idprovider: string;
  militaryStatus: string;
}